chore(msc3911): Some test cleanups
